+++ test/openBCIGanglion-test.js
@@ -184,21 +184,25 @@ describe('#ganglion', function () {
     let funcSpyCompressedData;
     let funcSpyUncompressedData;
     let funcSpyWrite;
+    let funcSpyDroppedPacket;
     before(function () {
       funcSpyWrite = sinon.spy(ganglion, 'write');
       funcSpyCompressedData = sinon.spy(ganglion, '_processCompressedData');
       funcSpyUncompressedData = sinon.spy(ganglion, '_processUncompressedData');
+      funcSpyDroppedPacket = sinon.spy(ganglion, '_droppedPacket');
     });
     beforeEach(function () {
-      funcSpyWrite.reset();
       funcSpyCompressedData.reset();
+      funcSpyDroppedPacket.reset();
       funcSpyUncompressedData.reset();
+      funcSpyWrite.reset();
       ganglion._resetDroppedPacketSystem();
     });
     it('should call proper functions if no dropped packets', function () {
       it('should work on uncompressed data', function () {
         ganglion._processProcessSampleData(utils.sampleUncompressedData());
         funcSpyUncompressedData.should.have.been.called;
+        funcSpyDroppedPacket.should.not.have.been.called;
       });
 
       it('should work on compressed data', function () {
@@ -209,6 +213,12 @@ describe('#ganglion', function () {
     it('should try to resend 0 packet and add packet 1 to buffer', function () {
       // Send the last buffer, set's ganglion._packetCounter
       ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdSampleMax));
+        funcSpyDroppedPacket.should.not.have.been.called;
+      });
+    it('should recognize 0 packet dropped', function () {
+      // Send the last buffer, set's ganglion._packetCounter
+      ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdSampleMax));
+      funcSpyCompressedData.should.have.been.called;
       const expectedMissedSample = k.OBCIGanglionByteIdUncompressed;
       // Call the function under test with one more then expected
       const nextPacket = utils.sampleCompressedData(expectedMissedSample + 1);
@@ -231,6 +241,8 @@ describe('#ganglion', function () {
       expect(ganglion._packetBuffer).to.deep.equal([]);
       expect(ganglion._requestedPacketResend).to.deep.equal([]);
 
+      funcSpyCompressedData.should.have.been.calledTwice;
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample);
     });
     it('should not find a dropped packet on wrap around', function () {
       ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdSampleMax - 1));
@@ -243,6 +255,7 @@ describe('#ganglion', function () {
       ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdUncompressed + 1));
       funcSpyCompressedData.should.have.been.calledThrice;
       funcSpyWrite.should.not.have.been.called;
+      funcSpyDroppedPacket.should.not.have.been.called;
     });
     it('should try to resend packet 127', function () {
       ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdSampleMax - 1));
@@ -253,6 +266,7 @@ describe('#ganglion', function () {
       expect(bufferEqual(funcSpyWrite.firstCall.args[0], (new Buffer([k.OBCIMiscResend, expectedMissedSample])))).to.equal(true);
       expect(ganglion._requestedPacketResend[0]).to.equal(expectedMissedSample);
       expect(bufferEqual(ganglion._packetBuffer[0], nextPacket)).to.equal(true);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample);
     });
     it('should try to resend packet 126 and 127', function () {
       ganglion._processProcessSampleData(utils.sampleCompressedData(k.OBCIGanglionByteIdSampleMax - 2));
@@ -265,6 +279,8 @@ describe('#ganglion', function () {
       expect(bufferEqual(funcSpyWrite.secondCall.args[0], (new Buffer([k.OBCIMiscResend, expectedMissedSample2])))).to.equal(true);
       expect(ganglion._requestedPacketResend).to.deep.equal([expectedMissedSample1, expectedMissedSample2]);
       expect(bufferEqual(ganglion._packetBuffer[0], nextPacket)).to.equal(true);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample1);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample2);
     });
     it('should try to resend packet 1 and add packet 2 to buffer', function () {
       // Send the last buffer, set's ganglion._packetCounter
@@ -276,6 +292,7 @@ describe('#ganglion', function () {
       expect(bufferEqual(funcSpyWrite.firstCall.args[0], (new Buffer([k.OBCIMiscResend, expectedMissedSample])))).to.equal(true);
       expect(ganglion._requestedPacketResend[0]).to.equal(expectedMissedSample);
       expect(bufferEqual(ganglion._packetBuffer[0], nextPacket)).to.equal(true);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample);
     });
     it('should try to resend packet 1 & 2 and add packet 3 to buffer', function () {
       // Send the last buffer, set's ganglion._packetCounter
@@ -289,6 +306,8 @@ describe('#ganglion', function () {
       expect(bufferEqual(funcSpyWrite.secondCall.args[0], (new Buffer([k.OBCIMiscResend, expectedMissedSample2])))).to.equal(true);
       expect(ganglion._requestedPacketResend).to.deep.equal([expectedMissedSample1, expectedMissedSample2]);
       expect(bufferEqual(ganglion._packetBuffer[0], nextPacket)).to.equal(true);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample1);
+      funcSpyDroppedPacket.should.have.been.calledWith(expectedMissedSample2);
     });
   });
   describe('_processBytes', function () {
@@ -441,9 +460,7 @@ describe('#ganglion', function () {
     // Makes sure the correct amount of events were called.
     expect(runningEventCount).to.equal(totalEvents);
   });
-
 });
-
 xdescribe('#noble', function () {
   xdescribe('#_nobleInit', function () {
     it('should emit powered on', function (done) {