+++ changelog.md
@@ -5,6 +5,7 @@
 
 ### Enhancements
 * Compress with 18bits vs 19bits
+* Reworked dropped packet detection and emit of packets.
 
 ### Bug Fixes
 * Fix bug where ganglionServer example would hang on scan if no board found.